Refactor hssp with cache
